pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bods-gql"
version = "0.1.0"
description = "Convert BODS v0.4 beneficial ownership data for querying with GQL (ISO/IEC 39075) on BigQuery"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Stephen Abbott Pugh"},
]
keywords = ["bods", "beneficial-ownership", "gql", "graph-query-language", "bigquery", "property-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]
dependencies = [
    "click>=8.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-bigquery-storage>=2.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
bods-gql = "bods_gql.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
